requirements.txt
django
django-ninja
PyJWT
whitenoise
dj-database-url
psycopg2-binary
boto3
python-dotenv
gunicorn
